+++ modules/admin/src/folder/Query.php
@@ -10,7 +10,7 @@ use luya\admin\storage\QueryTrait;
  *
  * See the {{\luya\admin\storage\QueryTrait}} for more informations.
  *
- * @property \luya\admin\components\StorageContainer $storage The storage component
+ * @property \luya\admin\storage\BaseFileSystemStorage $storage The storage component
  *
  * @author Basil Suter <basil@nadar.io>
  * @since 1.0.0
@@ -24,7 +24,7 @@ class Query extends \yii\base\Object
     /**
      * Singleton behavior for storage component getter.
      *
-     * @return \luya\admin\components\StorageContainer
+     * @return \luya\admin\storage\BaseFileSystemStorage
      */
     public function getStorage()
     {